finish ex4.15.-4.23

+++ part4/blogList/controllers/blogs.js
@@ -1,28 +1,46 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
+  const user = request.user
+  const userhandle = await User.findById(user.id)
+  // console.log('user in blogsRounter:', user)
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes,
+    user: userhandle._id
   })
 
   const savedBlog = await blog.save()
+  userhandle.blogs = userhandle.blogs.concat(savedBlog._id)
+  await userhandle.save()
   response.status(201).json(savedBlog)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(401).json({ error: 'cannot find a blog with that id' })
+  } else {
+    if (blog.user.toString() !== user.id.toString()) {
+      return response.status(401).json({ error: 'can only delete blog(s) of your own' })
+    } else {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    }
+  }
 })
 
 blogsRouter.put('/:id', (request, response, next) => {
